comments + refactor
