+++ Benchmarks/BarsDictionaryBenchmarkData.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using TradingStrategies.Backtesting.Optimizers;
+using TradingStrategies.Backtesting.Optimizers.Utility;
 using TradingStrategies.Utilities;
 using WealthLab;
